Fix checks for first and last row and go back to manual flushing
When drawing using multiple draw calls, the multiple flushes in the middle of drawing led to weird flickering.
It might actually be a better idea to buffer it into a `Vec<u8>` and then write it all out in `flush`, to guarantee that it doesn't get written out halfway if the `BufWriter` runs out of space.