Use a custom color type rather than `Rgb888`
It's basically the same as `crossterm`'s color, with one major difference - instead of `Color::Reset`, there's `Color::BgColor` and `Color::FgColor`.
Which color `Color::Reset` represents depends on whether it's being applied to the foreground or the background - in which cases it represents the default foreground color and the default background color respectively. So, I replaced it with the two actual colors it can generate, `FgColor` and `BgColor`, and added the logic necessary to make the correct halves into the foreground and background.

This also allows accessing the user-defined terminal colors.